refactor(pubsub): make exactly once retry loop logic a function
